Remove NotStarted execution phase
