+++ dnsconfd/fsm/transitions/__init__.py
@@ -1,5 +1,4 @@
 from .transition_implementations import TransitionImplementations
-from .not_started import NotStarted
 from .starting import Starting
 from .running import Running
 from .stopping import Stopping
